Add different tasks for main()
