+++ tasks/predict.py
@@ -1,0 +1,2 @@
+def predict(args):
+    ...
